requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
langchain==0.0.335
langchain-google-genai==0.0.5
google-generativeai==0.3.1
aiohttp==3.8.6
python-multipart==0.0.6
beautifulsoup4==4.12.2
tweepy==4.14.0
asyncpraw==7.7.1
arxiv==1.4.8
numpy==1.26.0
asyncio>=3.4.3
uuid>=1.30 